[Update] 주말 플랜에 대한 skill 이 생성되지 않는 문제 수정
