+++ src/main/resources/application-dev.properties
@@ -9,7 +9,7 @@ spring.datasource.password=1234
 
 spring.h2.console.enabled=true
 spring.jpa.open-in-view=false
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.default_batch_fetch_size=100
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
